Create Jadoo_and_strings.py
Jadoo's best friend is in trouble. In order to save him, he has been given a task. The task is : Print the reverse of a given String Input. But, as you know that Jadoo is too lazy, he decides to do this using shortest code possible. Please help him, will you?

INPUT

A given String S of length N [1<=N<=1000]

OUTPUT

The reverse of that whole string